+++ src/domain/payment/Invoice.java
@@ -1,13 +1,15 @@
 package domain.payment;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
+
 import domain.Entity;
 import domain.user.Subscriber;
 
 public class Invoice extends Entity<Long> {
     private Subscriber subscriber;
     private Timestamp invoicingDate;
-    private Float amount;
+    private BigDecimal amount;
     
     public Invoice() { }
 
@@ -27,11 +29,11 @@ public class Invoice extends Entity<Long> {
         this.invoicingDate = invoicingDate;
     }
 
-    public Float getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Float amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
